float to int in DrawableUtils

+++ app/src/main/java/com/seiko/demo/utils/ImageLoader.kt
@@ -25,7 +25,7 @@ class ImageLoader {
             if (radius == 0) {
                 view.setBackgroundColor(color)
             } else {
-                view.background = DrawableUtils.createRadiusDrawable(color, radius.toFloat())
+                view.background = DrawableUtils.createRadiusDrawable(color, radius)
             }
         }
